Use `float` consistently for the coordinates of users

+++ turms-server-common/src/main/java/im/turms/server/common/domain/location/bo/NearbyUser.java
@@ -26,8 +26,8 @@ import im.turms.server.common.domain.user.po.User;
 public record NearbyUser(
         Long userId,
         DeviceType deviceType,
-        Double longitude,
-        Double latitude,
+        Float longitude,
+        Float latitude,
         Integer distance,
         User info
 ) {
